Rescale to output type range
When writing into disk (voxelize tool) and the output data type
is different than float, rescale according to output data type.
The same rescale operation is applied in the Livre data source,
to rescale from float to uint8_t.
All the rescale/cast operations are done at the end of the
workflow, not in the loaders/functors. Tests adapted accordingly.