Fix the volume writing when there are dots in the output path
When a "." was present in the specified --output path, the voxelize
app was failing to find the extension, leading to unexpected behaviors.
Now only the file name is considered to search for the extension.
Also, now it's possible to write also series of volumes using a different
format than the default (.mhd).